+++ pennylane_lightning/core/src/simulators/lightning_qubit/StateVectorLQubit.hpp
@@ -241,6 +241,29 @@ class StateVectorLQubit : public StateVectorBase<PrecisionT, Derived> {
     }
 
     /**
+     * @brief Apply a single gate to the state-vector.
+     *
+     * @param opName Name of gate to apply.
+     * @param wires Wires to apply gate to.
+     * @param inverse Indicates whether to use inverse of gate.
+     * @param params Optional parameter list for parametric gates.
+     * @param matrix Matrix data (in row-major format).
+     */
+    template <typename Alloc>
+    void applyOperation(
+        [[maybe_unused]] const std::string &opName,
+        const std::vector<size_t> &wires, bool inverse,
+        const std::vector<PrecisionT> &params,
+        [[maybe_unused]] const std::vector<ComplexT, Alloc> &matrix) {
+        auto &dispatcher = DynamicDispatcher<PrecisionT>::getInstance();
+        if (dispatcher.hasGateOp(opName)) {
+            applyOperation(opName, wires, inverse, params);
+        } else {
+            applyMatrix(matrix, wires, inverse);
+        }
+    }
+
+    /**
      * @brief Apply a single generator to the state-vector using a given kernel.
      *
      * @param kernel Kernel to run the operation.